#3 - 주석 수정
변경 사항
- 포트폴리오 정렬 함수 파일로 분리
